Fix build error due to restructuring of Ballerina core dispathing packages
