+++ modules/services/workspace-service/src/main/java/org/wso2/ballerina/tooling/service/workspace/swagger/SwaggerResourceMapper.java
@@ -21,7 +21,7 @@ import io.swagger.models.Path;
 import io.swagger.models.Response;
 import org.wso2.ballerina.core.model.Annotation;
 import org.wso2.ballerina.core.model.Resource;
-import org.wso2.ballerina.core.runtime.dispatching.Constants;
+import org.wso2.ballerina.core.runtime.dispatching.http.Constants;
 
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
